% Generated by roxygen2 (4.0.1): do not edit by hand
\name{oaxaca}
\alias{oaxaca}
\title{Oaxaca Decomposition for linear models}
\usage{
oaxaca(formula, data, group, detailed = FALSE, decomposition = "oaxaca_0")
}
\arguments{
\item{group}{The string character for the column in the \code{data} which identifies different groups.
The group variable must have at least 2 unique values.}

\item{decomposition}{Which decomposition to use. Possible values are \code{oaxaca_0}, \code{oaxaca_1},
\code{Cotton}, \code{Reimer} and \code{Neumark}. For more information on the meaning of these decompositions
see '\code{Details}'.}

\item{detailed}{If \code{TRUE}, prints the summaries of all models on top of decomposition summaries}
}
\description{
The function creates a seperate linear regression for each level in the \code{group}
 and apply the Blinder-Oaxaca Decomposition technique.
}
\details{
The gap between the predicted means of different groups will always be calculated so that
 the gap is positive. For the moment, the function cannot show the interactions between the
 endowment gap (explained) and the coefficient gap (unexplained). The function can only handle up to
 decomposition for 3 groups.
}
\author{
Christopher Lee
}

